[refactor] user service method extraction.
