+++ src/main/java/com/prgrms/needit/domain/user/login/service/UserService.java
@@ -40,30 +40,6 @@ public class UserService {
 		this.authManagerBuilder = authManagerBuilder;
 	}
 
-	public Long getCurUserId() {
-		if(getCurCenter().isPresent() && getCurMember().isEmpty()) {
-			return getCurCenter().get().getId();
-		}
-
-		if(getCurMember().isPresent() && getCurCenter().isEmpty()) {
-			return getCurMember().get().getId();
-		}
-
-		throw new NotFoundResourceException(ErrorCode.NOT_FOUND_USER);
-	}
-
-	public UserType getCurUserType() {
-		if(getCurCenter().isPresent() && getCurMember().isEmpty()) {
-			return UserType.CENTER;
-		}
-
-		if(getCurMember().isPresent() && getCurCenter().isEmpty()) {
-			return UserType.MEMBER;
-		}
-
-		throw new NotFoundResourceException(ErrorCode.NOT_FOUND_USER);
-	}
-
 	public TokenResponse login(LoginRequest login) {
 		findActiveMemberAndCenter(login.getEmail());
